refactor(rental): move RentCarDTO to rental/dtos dir
