+++ src/modules/rental/useCases/rentCar/RentCarController.ts
@@ -3,7 +3,7 @@ import { container } from 'tsyringe';
 import { AppError } from '@shared/errors/AppError';
 import { dataValidation } from '@shared/utils/dataValidation';
 import { RentCarUseCase } from './RentCarUseCase';
-import { RentCarDTO } from './RentCarDTO';
+import { RentCarDTO } from '../../dtos/RentCarDTO';
 
 export class RentCarController {
   async handle(request: Request, response: Response): Promise<Response> {
